+++ constants.h
@@ -106,9 +106,9 @@
 
 typedef struct
 {
-    unsigned int function;
-    unsigned int maxInputs;
-    unsigned int inputs[MAX_ARITY];
+    unsigned short int function;
+    unsigned short int maxInputs;
+    unsigned short int inputs[MAX_ARITY];
     float inputsWeight[MAX_ARITY];
     int active;
 } Node;
@@ -116,57 +116,57 @@ typedef struct
 typedef struct
 {
     Node nodes[MAX_NODES];
-    unsigned int output[MAX_OUTPUTS];
-    unsigned int activeNodes[MAX_NODES];
-    unsigned int numActiveNodes;
+    unsigned short int output[MAX_OUTPUTS];
+    unsigned short int activeNodes[MAX_NODES];
+    unsigned short int numActiveNodes;
     float fitness;
     float fitnessValidation;
 } Chromosome;
 
 typedef struct
 {
-    unsigned int function_inputs_active;
-    unsigned int inputs[MAX_ARITY/2];
+    unsigned short int function_inputs_active;
+    unsigned short int inputs[MAX_ARITY/2];
     float inputsWeight[MAX_ARITY];
 } CompactNode;
 
 typedef struct
 {
     CompactNode nodes[MAX_NODES];
-    unsigned int output[MAX_OUTPUTS];
-    unsigned int activeNodes[MAX_NODES/2];
-    unsigned int numActiveNodes;
+    unsigned short int output[MAX_OUTPUTS];
+    unsigned short int activeNodes[MAX_NODES/2];
+    unsigned short int numActiveNodes;
 } CompactChromosome;
 
 
 typedef struct
 {
-    unsigned int function;
-    unsigned int maxInputs;
-    unsigned int inputs[MAX_ARITY];
+    unsigned short int function;
+    unsigned short int maxInputs;
+    unsigned short int inputs[MAX_ARITY];
     float inputsWeight[MAX_ARITY];
-    unsigned int originalIndex;
+    unsigned short int originalIndex;
 
 } ActiveNode;
 
 typedef struct
 {
     ActiveNode nodes[MAX_NODES];
-    unsigned int output[MAX_OUTPUTS];
-    unsigned int numActiveNodes;
+    unsigned short int output[MAX_OUTPUTS];
+    unsigned short int numActiveNodes;
 } ActiveChromosome;
 
 
 typedef struct
 {
-    unsigned int N; //inputs
-    unsigned int O; //outputs
-    unsigned int M; //dataset size
+    unsigned short int N; //inputs
+    unsigned short int O; //outputs
+    unsigned short int M; //dataset size
 
     unsigned short int NUM_FUNCTIONS;
     unsigned short int* functionSet;
-   // unsigned int* maxFunctionInputs;
-    //unsigned int* inputVariablesSet;
+   // unsigned short int* maxFunctionInputs;
+    //unsigned short int* inputVariablesSet;
 
     float weightRange;
     char** labels;
@@ -175,11 +175,11 @@ typedef struct
 typedef struct
 {
     /** Number of inputs */
-    unsigned int N;
+    unsigned short int N;
     /** Number of outputs */
-    unsigned int O;
+    unsigned short int O;
     /** Number of entries */
-    unsigned int M;
+    unsigned short int M;
 
     unsigned short int** data;
     unsigned short int** output;
